Big bunch of new graphics for the standard png set.
Especially noticeable is the new picture for
the chinese dragon. Due to the special "importance"
of this dragon I had set up a poll weither the
image is appreciate, which turned out 7:3 in favour.
Note that the old image continues to exist both
in the xpm and alternate png set.

--AndreasV


git-svn-id: https://svn.code.sf.net/p/crossfire/code/trunk/arch@1274 282e977c-c81d-0410-88c4-b93c2d0d6712